Removed scala-react from the project, and replace it with a binary.
The reason for the change is that many people got confused over whether
scala-react is part of the dissertation work (answer: it is not).
In order to clear things up, I replaced it with a local repo that
contains it compiled into a .jar file.

Maybe a nicer solution would be to upload it to maven central or
equivalent.